+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [13.21.1](https://kaos.sh/ek/13.21.1)
 
 - **`[timeutil]`** Added method `LocalTimezone`
+- **`[req]`** Added helper `Response.Save`
 
 ### [13.21.0](https://kaos.sh/ek/13.21.0)